Module aws_sdk_iotsecuretunneling::model

Modules

connection_state

See ConnectionState

destination_config

See DestinationConfig

tag

See Tag

timeout_config

See TimeoutConfig

tunnel

See Tunnel

tunnel_summary

See TunnelSummary

Structs

ConnectionState

The state of a connection.

DestinationConfig

The destination configuration.

Tag

An arbitary key/value pair used to add searchable metadata to secure tunnel

TimeoutConfig

Tunnel timeout configuration.

Tunnel

A connection between a source computer and a destination device.

TunnelSummary

Information about the tunnel.

Enums

ConnectionStatus

TunnelStatus